Address PHP snippet review feedback
